Code cleanup

+++ src/main/java/net/openhft/chronicle/engine/map/CMap2EngineReplicator.java
@@ -1,17 +1,19 @@
 /*
- *     Copyright (C) 2015  higherfrequencytrading.com
  *
- *     This program is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU Lesser General Public License as published by
- *     the Free Software Foundation, either version 3 of the License.
+ *  *     Copyright (C) ${YEAR}  higherfrequencytrading.com
+ *  *
+ *  *     This program is free software: you can redistribute it and/or modify
+ *  *     it under the terms of the GNU Lesser General Public License as published by
+ *  *     the Free Software Foundation, either version 3 of the License.
+ *  *
+ *  *     This program is distributed in the hope that it will be useful,
+ *  *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  *     GNU Lesser General Public License for more details.
+ *  *
+ *  *     You should have received a copy of the GNU Lesser General Public License
+ *  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     GNU Lesser General Public License for more details.
- *
- *     You should have received a copy of the GNU Lesser General Public License
- *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 package net.openhft.chronicle.engine.map;
@@ -96,7 +98,7 @@ public class CMap2EngineReplicator implements EngineReplication,
         for (long i = b.start(); i < b.readLimit(); i++) {
             int ch = b.readByte(i);
             if (ch < ' ')
-                throw new AssertionError("Char " + (int) ch);
+                throw new AssertionError("Char " + ch);
         }
     }
 
@@ -108,13 +110,13 @@ public class CMap2EngineReplicator implements EngineReplication,
         for (long i = 0; i < 16 && i < b.limit(); i++) {
             int ch = b.readByte(i);
             if (ch < ' ')
-                throw new AssertionError("Char " + (int) ch);
+                throw new AssertionError("Char " + ch);
         }
         if (b.limit() > 32)
             for (long i = b.limit() - 16; i < b.limit(); i++) {
                 int ch = b.readByte(i);
                 if (ch < ' ')
-                    throw new AssertionError("Char " + (int) ch);
+                    throw new AssertionError("Char " + ch);
             }
     }
